Add Two More Recipes and Optimize the Webpages
The recipes for macaroni salad and chicken adobo was added to complete the
full course meal recipes. These were contained in macaroni.html and
adobo.html. The return to homepage button was also added in each recipe pages
so that the user can go back and forth with them. Lastly, the homepage was
redesigned to make it more appealing.